Add files via upload
Each book in the library has a book code, which can be used for quick book retrieval. This book code is a positive integer.

Each reader who borrows books has a demand code in hand, and this demand code is also a positive integer. If the book code of a book ends exactly with the reader's demand code, then this book is exactly what the reader needs.

 D has just become the librarian. She knows the book codes of all the books in the library. She asks you to help her write a program to find the book with the smallest book code among the books that each reader needs. If there is no book he needs, please output -1.