Add files via upload
I received a strange note: "Let's go on a date!" " 3485djDkxh4hhGE  2984akDfkkkkggEdsb s&hgsfdk d&Hyscvnm" The great detective soon understood that the strange garbled characters on the note were actually the appointment time "Thursday 14:04", because the first pair of the same capital English letters in the first two strings (with a distinction between upper and lower cases) was the fourth letter 'D', representing Thursday. The second pair of identical characters is 'E', which is the fifth letter of the English alphabet and represents the 14th hour of the day (thus, from 0:00 to 23:00 of the day is represented by the numbers 0 to 9 and the capital letters A to N). The first pair of the same English letters 's' in the following two strings appears at the 4th position (counting from 0), representing the 4th minute. Given two pairs of strings, please help Holmes decode them to obtain the time of the appointment.



(Note: The comparison process of two strings must be carried out based on the characters with the same index corresponding to the two strings!)"