Add files via upload
A census was conducted in a certain town, and the birthdays of all the residents were obtained. Now please write a program to find out the oldest and youngest people in the town.







Here, it is ensured that each input date is legal, but not necessarily reasonable - suppose it is known that there are no people over 200 years old in the town, and today is September 6, 2014, so it exceeds 200



Both the birthday of the current year and the birthday before birth are unreasonable and should be filtered out.