Add files via upload
A few keys on the old keyboard were broken, so when typing a piece of text, the corresponding characters would not appear. Now, please list the text that should be input and the actual text that has been input

Those keys must be broken.



Input description:

Input the text that should be input and the actual text that is input are respectively provided in two lines. Each text is A string of no more than 80 characters, consisting of the letters A-Z (including both uppercase and lowercase), and the numbers 0-9.

It is composed of an underscore "_" (representing a space). The problem guarantees that neither of the two strings is empty.





Output description

Output the broken keys in one line in the order of discovery. Among them, only uppercase English letters are output, and each bad key is output only once. The problem must have at least one bad key.