Add files via upload
Title description

Given a sequence of positive integers and a positive integer p, let the maximum value in this sequence be M and the minimum value be M. If M <= m * p, then this sequence is called a perfect sequence.







Now, given the parameter p and some positive integers, please select as many numbers as possible from them to form a perfect sequence.



Input description:

The first line of the input provides two positive integers N and p, where N (<= 10 ^ 5) is the number of positive integers input and p (<= 10 ^ 9) is the given parameter. The second line gives N positive integers, each number

No more than 109.





Output description

Output in one line the maximum number of numbers that can be selected to form a perfect sequence with them.