Add files via upload
A legal ID number consists of a 17-digit region code, date code, sequence code and a 1-digit check code. The calculation rule of the check code is as follows:

First, calculate the weighted sum of the first 17 digits, and the weight distribution is: {7,9,10,5,8,4,2,1,6,3,7,9,10,5,8,4,2}; Then take the modulus of the calculated sum to 11

To the value Z; Finally, match the Z value with the check code M according to the following relationship:

Z: 0 1 2 3 4 5 6 7 8 9 10

M: 1 0 X 9 8 7 6 5 4 3 2

Now, given some ID card numbers, please verify the validity of the check code and output the problematic numbers.