Add files via upload
Input description:

The input should provide a positive integer N not exceeding 105 on the first line, which is the number of participants. The following N lines each provide the information and score of one participant, including the number of the school they represent and their competition score (out of 100), separated by Spaces.





Output description

In one line, provide the number of the school with the highest total score and its total score, separated by Spaces. The answers to the questions are guaranteed to be unique and there are no parallel answers.