Add files via upload
Title description

A few keys on the old keyboard were broken, so when typing a piece of text, the corresponding characters would not appear. Now, given the text that should be input and the broken keys, what would the resulting text be like?



Input description:

Input the broken keys and the text that should be input in two lines respectively. The bad keys corresponding to the English letters are given in uppercase. Each paragraph is a string of no more than 10^5 characters. The available characters include the letters [a-z, A-Z], the numbers 0-9, and the underscores "_" (representing Spaces), ", ", "." "-", "+" (representing the up gear key). The question ensures that the string of text input in the second line is not empty.

Note: If the up key is broken, capital English letters cannot be typed.





Output description

Output the result text that can be typed in one line. If no character can be typed, output an empty line.