Add files via upload
Title description

Xiaohong wants to buy some beads to make a string of beads she likes. The vendor selling beads had many strings of colorful beads, but refused to take any one of them apart for sale. So Xiaohong asks you to help judge one

Next, does a certain string of beads contain all the beads you want? If so, then tell her how many extra beads there are. If not, then tell her how many beads are missing.



For convenience, we use characters within the range of [0-9], [a-z], and [A-Z] to represent colors. For instance, YrR8RrY is the bead string that Xiaohong wants to make. Then ppRYYGrrYBR2258 can be purchased because it includes it

All the beads she wanted, and there were eight more that she didn't need. ppRYYGrrYB225 cannot be purchased because it lacks black beads and one red bead is missing.



Input description:

Each input contains one test case. Each test case successively presents the beaded string of the stall owner and the beaded string that Xiaohong wants to make in two lines, with no more than 1,000 beads in each string.





Output description

If it can be purchased, output "Yes" in one line and how many extra beads there are. If it cannot be purchased, output "No" and the number of missing beads in one line. It is separated by one space in between.