Add files via upload
Title description

There is a cow and it gives birth to a heifer at the beginning of each year. Each heifer gives birth to a heifer at the beginning of each year starting from the fourth year. Please program to realize how many cows there are in total in the NTH year?



Input description:

The input data consists of multiple test cases, each occupying one line, including an integer n(0<n<55), the meaning of which is described in the question.





Output description

For each test case, output the number of cows in the NTH year.

Each output should be on one line.