Add files via upload
NowCoder always strives to be the best and always wants to be the first in everything, so he has a special fondness for the number "1". By extension of love, he also likes numbers containing 1, such as 10, 11, 12... . Not only that, he also likes numbers that contain 1 in any base system. Of course, the more 1s it contains, the more he likes it. Could you help him count the number of 1s of a certain number in a specific base system?



Input description:

The input contains multiple sets of data, each of which consists of a positive integer n (1≤n≤2147483647) and a positive integer r (2≤r≤16).

Here, n is a decimal integer and r is the target base.





Output description

For each input group, after the output integer n is converted to base r, how many "1s" are contained therein?